experiments/agent-scoped-tools-triage/requirements.txt
requests==2.33.0
pyjwt==2.12.1
